move state to mod
